// apps/learning-analytics-web/src/app/main/report/enums/report.ts
export enum ReportType {
  WORKSPACE_OVERVIEW = 'WORKSPACE_OVERVIEW',
  COURSE_OVERVIEW = 'COURSE_OVERVIEW',
  USER_OVERVIEW = 'USER_OVERVIEW',
  WORKSPACE_MISSION = 'WORKSPACE_MISSION',
  MISSION_ENROLLMENTS = 'MISSION_ENROLLMENTS',
  MISSION_ENROLLMENTS_QUIZZES = 'MISSION_ENROLLMENTS_QUIZZES',
  MISSION_QUIZ = 'MISSION_QUIZ',
  GROUP_MISSION_USER = 'GROUP_MISSION_USER',
  MISSION_EVALUATIONS = 'MISSION_EVALUATIONS',
  MISSION_EVALUATION_ANALYSIS = 'MISSION_EVALUATION_ANALYSIS',
  PULSE_CHANNELS = 'PULSE_CHANNELS',
  PULSES_ACTIVITIES = 'PULSES_ACTIVITIES',
  PULSES_QUIZ = 'PULSES_QUIZ',
  GROUP_CHANNEL_USER = 'GROUP_CHANNEL_USER',
  ALL_USERS = 'ALL_USERS',
  USERS_ACCESS = 'USERS_ACCESS',
  USER_GENERAL_STATISTICS = 'USER_GENERAL_STATISTICS',
  USERS_GENERAL_CONSUMPTION = 'USERS_GENERAL_CONSUMPTION',
  SMARTZAP_COURSE_OVERVIEW = 'SMATZAP_COURSE_OVERVIEW',
  TRAIL_LIST = 'TRAIL_LIST',
  TRAIL_ENROLLMENTS = 'TRAIL_ENROLLMENTS',
  TRAIL_CONCLUSION_RATE = 'TRAIL_CONCLUSION_RATE',
  USER_PERMISSIONS = 'USER_PERMISSIONS',
}